Postgresql/app-postgresql-main/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
Flask-Ext==0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
requests==2.31.0
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
pip==23.3.1
setuptools==68.2.2